moved timezones and locations into dedicated files
